requirements-ai.txt
aiofile==3.9.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
authlib==1.6.9
bandit==1.9.4
beartype==0.22.9
boto3==1.42.73
botocore==1.42.81
cachetools==7.0.5
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
cohere==5.20.7
cryptography==46.0.6
cyclopts==4.10.1
deprecation==2.1.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fastavro==1.12.1
fastmcp==3.1.1
fastuuid==0.14.0
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
feedparser>=6.0.11
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.8.0
idna==3.11
importlib-metadata==8.7.1
iniconfig==2.3.0
jaraco-classes==3.4.0
jaraco-context==6.1.2
jaraco-functools==4.4.0
jeepney==0.9.0
jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
lance-namespace==0.6.1
lance-namespace-urllib3-client==0.6.1
lancedb==0.29.2
litellm==1.82.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.27.0
mdurl==0.1.2
more-itertools==11.0.1
multidict==6.7.1
numpy==2.4.4
ollama==0.6.1
openai==2.28.0
openapi-pydantic==0.5.1
opentelemetry-api==1.40.0
packaging==26.0
pandas==3.0.2
pathable==0.5.0
pip-audit==2.9.0
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
propcache==0.4.1
py-key-value-aio==0.4.4
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pygments==2.20.0
pyjwt==2.12.1
pyperclip==1.11.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-timeout==0.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pyyaml==6.0.3
referencing==0.37.0
regex==2026.3.32
requests==2.33.0
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
ruff==0.15.6
s3transfer==0.16.0
secretstorage==3.5.0
sentry-sdk==2.57.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sse-starlette==3.3.4
starlette==0.52.1
stevedore==5.7.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
types-requests==2.33.0.20260402
typing-extensions==4.15.0
typing-inspection==0.4.2
uncalled-for==0.2.0
urllib3==2.6.3
uvicorn==0.42.0
vt-py==0.22.0
watchfiles==1.1.1
websockets==16.0
yarl==1.23.0
zipp==3.23.0
